refactor: 유기동물 api 모듈화
